Update script.js
A couple of minor changes, along with making it so moving between solves of an average doesn't add to the history. This makes it so the back arrow will bring you to the home page instead of needing to press it several times even if you thought you didn't go to any new urls.